change impl to explicitly store length as u32, rather than size_t, so we keep a packed struct on 64bit machines
BUG=

Review URL: https://codereview.chromium.org/16356006

git-svn-id: http://skia.googlecode.com/svn/trunk@9436 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkString.h
@@ -205,7 +205,7 @@ public:
 private:
     struct Rec {
     public:
-        size_t      fLength;
+        uint32_t    fLength; // logically size_t, but we want it to stay 32bits
         int32_t     fRefCnt;
         char        fBeginningOfData;
